기능추가 : Milestone 도메인 패키지, 엔티티 등 구조 작성
