--- src/main/java/kr/co/co_working/common/entity/CommonTime.java
@@ -1,24 +1,0 @@
-package kr.co.co_working.common.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class CommonTime {
-        @CreatedDate
-        @Column(name = "created_at")
-        private LocalDateTime createdAt;
-
-        @LastModifiedDate
-        @Column(name = "modified_at")
-        private LocalDateTime modifiedAt;
-}
