+++ src/main/java/kr/co/co_working/member/Member.java
@@ -1,8 +1,8 @@
-package kr.co.co_working.member.repository.entity;
+package kr.co.co_working.member;
 
 import jakarta.persistence.*;
-import kr.co.co_working.common.entity.CommonTime;
-import kr.co.co_working.memberTeam.repository.entity.MemberTeam;
+import kr.co.co_working.common.CommonTime;
+import kr.co.co_working.memberTeam.MemberTeam;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
