+++ src/main/java/kr/co/co_working/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package kr.co.co_working.member.repository;
 
-import kr.co.co_working.member.repository.entity.Member;
+import kr.co.co_working.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, String> {
