+++ src/main/java/kr/co/co_working/memberTeam/repository/MemberTeamRepository.java
@@ -1,6 +1,6 @@
 package kr.co.co_working.memberTeam.repository;
 
-import kr.co.co_working.memberTeam.repository.entity.MemberTeam;
+import kr.co.co_working.memberTeam.MemberTeam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberTeamRepository extends JpaRepository<MemberTeam, Long> {
