+++ src/main/java/kr/co/co_working/project/repository/ProjectRepository.java
@@ -1,6 +1,6 @@
 package kr.co.co_working.project.repository;
 
-import kr.co.co_working.project.repository.entity.Project;
+import kr.co.co_working.project.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProjectRepository extends JpaRepository<Project, Long> {
