+++ src/main/java/kr/co/co_working/project/service/ProjectService.java
@@ -4,7 +4,7 @@ import kr.co.co_working.project.dto.ProjectRequestDto;
 import kr.co.co_working.project.dto.ProjectResponseDto;
 import kr.co.co_working.project.repository.ProjectDslRepository;
 import kr.co.co_working.project.repository.ProjectRepository;
-import kr.co.co_working.project.repository.entity.Project;
+import kr.co.co_working.project.Project;
 import kr.co.co_working.team.repository.TeamRepository;
 import kr.co.co_working.team.repository.entity.Team;
 import lombok.RequiredArgsConstructor;
