+++ src/test/java/kr/co/co_working/member/service/MemberServiceTest.java
@@ -3,7 +3,7 @@ package kr.co.co_working.member.service;
 import kr.co.co_working.member.dto.MemberRequestDto;
 import kr.co.co_working.member.dto.MemberResponseDto;
 import kr.co.co_working.member.repository.MemberRepository;
-import kr.co.co_working.member.repository.entity.Member;
+import kr.co.co_working.member.Member;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
