+++ src/test/java/kr/co/co_working/task/service/TaskServiceTest.java
@@ -4,12 +4,12 @@ import kr.co.co_working.member.dto.MemberRequestDto;
 import kr.co.co_working.member.service.MemberService;
 import kr.co.co_working.project.dto.ProjectRequestDto;
 import kr.co.co_working.project.repository.ProjectRepository;
-import kr.co.co_working.project.repository.entity.Project;
+import kr.co.co_working.project.Project;
 import kr.co.co_working.project.service.ProjectService;
 import kr.co.co_working.task.dto.TaskRequestDto;
 import kr.co.co_working.task.dto.TaskResponseDto;
 import kr.co.co_working.task.repository.TaskRepository;
-import kr.co.co_working.task.repository.entity.Task;
+import kr.co.co_working.task.Task;
 import kr.co.co_working.team.dto.TeamRequestDto;
 import kr.co.co_working.team.service.TeamService;
 import org.junit.jupiter.api.Assertions;
